Function simple_parallel::for_

pub fn for_<I: IntoIterator, F>(iter: I, f: F) where I::Item: Send, F: Fn(I::Item) + Sync

Execute f on each element of iter, in their own scoped thread.

If f panics, so does for_. If this occurs, the number of elements of iter that have had f called on them is unspecified.