Function simple_parallel::unordered_map

pub fn unordered_map<'a, I: IntoIterator, F, T>(iter: I, f: &'a F) -> UnorderedParMap<'a, T> where Item: Send + 'a, F: 'a + Sync + Fn(Item) -> T, T: Send + 'static

Execute f on each element in iter, with unspecified yield order.

This behaves like simple_parallel::map, but does not make efforts to ensure that the elements are returned in the order of iter, hence this is cheaper.