Module simple_parallel::one_to_one

Structs

ParMap
A parallel-mapping iterator. Constructed by calling map.

UnorderedParMap
A parallel-mapping iterator that doesn't care about the order in which elements come out. Constructed by calling unordered_map.

Functions

map
Execute f on iter, yielding the results in the order the correspond to in iter.

unordered_map
Execute f on each element in iter, with unspecified yield order.