Module simple_parallel::pool

Structs

JobHandle
A token representing a job submitted to the thread pool.

ParMap
A parallel-mapping iterator, that yields elements in the order they are yielded by the underlying iterator. Constructed by calling Pool::map.

Pool
A thread pool.

UnorderedParMap
A parallel-mapping iterator, that yields elements in the order they are computed, not the order from which they are yielded by the underlying iterator. Constructed by calling Pool::unordered_map.